third_party/blink/web_tests/external/wpt/webrtc-extensions/RTCRtpReceiver-jitterBufferTarget-stats-expected.txt

This is a testharness.js-based test.
FAIL measure raising and lowering video jitterBufferTarget assert_equals: jitterBufferTarget supported for video expected (object) null but got (undefined) undefined
FAIL measure raising and lowering audio jitterBufferTarget assert_equals: jitterBufferTarget supported for audio expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
